Return response body if no processor matches (#7)
Previously only data handled by a `Processor` was made available to as a resource.

This commit changes the base processor to returned the response body in case of no processor accepting the payload. This makes it easier to to access e.g. binary file content. The patch includes new specs on the base processor.