Add a Msgpack processor
This is intended to use as a drop-in replacement for JSON. The
two encodings support the same datatypes, thus HTTP's content
negotiation can be used to tell a server to deliver a Msgpack
response (if it is capable to do so). Fully transparent to the
API consumer, this can reduce the HTTP payload by 50 percent
in some cases.

Example HTTP Request:

    GET /path HTTP/1.1
    Host: restify.de
    Accept: application/msgpack, application/json

If the server supports msgpack, it will deliver that
serialization; if it does not, it will fall back to JSON.
Restify will be able to handle both.

Spec: https://msgpack.org/
Library: https://github.com/msgpack/msgpack-ruby