Do not stop EM after each spec
Stopping is not synchronous and passes an reactor into the next spec that is
shutting down. Because the new spec already has requests scheduled and the
reactor has been reported as (still) running but does not actually process
requests the new specs fill hang indefinitely.

This commit changes the hook to stop EM only after all specs have been run.