Rebuild typhoeus adapter to use internal queue
This CL changes the internal implementation of the typhoeus adapter. It
adds an internal queue for new requests that replaced the previous
multi-thread access to Typhoeus::Hydra.

The new implementation adds new requests in an internal queue that is
only checked when either hydra finished or a request completes. At these
points, new requests are added to hydra. This has the downside that a
single long request can block newly added requests, that will not be
started before the long request has finished. It is recommended to use
short timeouts for all requests.

The advantages are a much simpler and actually correct implementation.
The queue provides proper thread synchronization, and hydra and the
libcurl FFI handles are not accessed from multiple threads anymore. This
avoids a few sporadically appearing race condition and threading bugs.
This should provide a much higher stability.