Regularly run minor GC in typhoeus background thread
When a client runs a lot of requests, somewhere in typhoeus/ethon/ffi, a
lot of memory gets allocated. It does not appear to reclaimed according
to memory profiling but also isn't release even on pressure.

This results in memory growing fast when doing a lot of requests that
isn't freed anymore and can result in the processing being killed by the
system due to OOM. That looks like a normal memory leak, but profiling
does not show any large amounts of reclaimed (!) memory. Also regularly
running a GC run on the background thread has visible effects on overall
long-running memory usage.

This CL adds a minor GC run every 100 hydra runs, if the queue is empty
(e.g. no immediate pending requests are present). This appears to have
little impact on performance but instantly keeps total memory
consumption in check.

This might get more configuration knobs and an on/off switch in future
releases.