fix: WebMock instrumentation with Typhoeus
The WebMock instrumentation for Typhoeus::Hydra didn't work anymore with
the patches Restify added to it's Hydra object to keep track of the OTEL
context. The patched removed the callback behavior in Hydra that was
used by WebMock.

This commit fixes that by keeping tracking of the OTEL context with the
Typhoeus request itself, restoring the context when Hydra actually adds
the request. Otherwise, the same code is called as before and callbacks
will work again too.

The WebMock support is further expanded to catch exception and bubble
them up to the promise, e.g. when request ain't allowed. Before, they
only stopped the background thread, and promises run into a timeout.
Now, whenever an exception is raised to a specific request, it will
instead reject the Promise of that request.

Lastly, the commit adds specs for the WebMock integration as well as the
working OTEL context propagation.