Extend test coverage for boolean query parameters
The previous implementation is handling expected query parameters with a `false` value in an unintended manner. Rather than appending the parameter as expected, it is omitted. This is in contrast to parameters not included in the pattern, where a `false` value would be appended.

In a test-driven development fashion, this commit adds a failing test first and thereby allows working on the bug more easily.